Hometask5/requirements.txt
flake8==3.7.1
pycodestyle==2.5.0
requests==2.22.0
coverage==4.5.4
nose==1.3.7
PyYAML==5.1.2

